Student Refactoring
