+++ backend/kaiser/src/main/java/com/example/springKaiser/SOLID/OpenClosedPrinciple/BetterFilter.java
@@ -1,6 +1,6 @@
 package com.example.springKaiser.SOLID.OpenClosedPrinciple;
 
-import com.example.springKaiser.entities.Students;
+import com.example.springKaiser.school.entities.Students;
 
 import java.util.List;
 import java.util.stream.Stream;
