+++ backend/kaiser/src/main/java/com/example/springKaiser/SOLID/OpenClosedPrinciple/NameSpecification.java
@@ -1,6 +1,6 @@
 package com.example.springKaiser.SOLID.OpenClosedPrinciple;
 
-import com.example.springKaiser.entities.Students;
+import com.example.springKaiser.school.entities.Students;
 
 public class NameSpecification implements Specification<Students> {
