+++ backend/kaiser/src/main/java/com/example/springKaiser/controller/SubscriberController.java
@@ -1,10 +1,8 @@
 package com.example.springKaiser.controller;
 
-import com.example.springKaiser.entities.Students;
 import com.example.springKaiser.entities.Subscriber;
 import com.example.springKaiser.repositories.SubscriberRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
