+++ backend/kaiser/src/main/java/com/example/springKaiser/school/Repository/GradeRepository.java
@@ -1,6 +1,6 @@
-package com.example.springKaiser.repositories;
+package com.example.springKaiser.school.Repository;
 
-import com.example.springKaiser.entities.Grades;
+import com.example.springKaiser.school.entities.Grades;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface GradeRepository  extends JpaRepository<Grades, Integer> {
