+++ backend/kaiser/src/main/java/com/example/springKaiser/school/Repository/TeacherRepository.java
@@ -1,6 +1,6 @@
-package com.example.springKaiser.repositories;
+package com.example.springKaiser.school.Repository;
 
-import com.example.springKaiser.entities.Teachers;
+import com.example.springKaiser.school.entities.Teachers;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TeacherRepository  extends JpaRepository<Teachers, Integer> {
