+++ backend/kaiser/src/main/java/com/example/springKaiser/school/entities/Students.java
@@ -1,4 +1,4 @@
-package com.example.springKaiser.entities;
+package com.example.springKaiser.school.entities;
 
 import jakarta.persistence.*;
 import lombok.Data;
